Fix variable type mismatch build failure

+++ runtime/core/array_ref.h
@@ -162,7 +162,7 @@ class ArrayRef final {
   /// slice(n, m) - Take M elements of the array starting at element N
   ArrayRef<T> slice(size_t N, size_t M) const {
     // cant slice longer then the array
-    size_t end = 0;
+    uint64_t end = 0;
     ET_CHECK(!c10::add_overflows(N, M, &end) && end <= size());
     return ArrayRef<T>(data() + N, M);
   }
